Fix issue where sharks are flung across screen if tab is hidden
`setInterval` will still run the callback if the tab is
in the background so the sharks' position would still
be updated. UI updates were deferred until the tab
received focus upon which the sharks would move several
iterations away from the previously visible position

This change uses `requestAnimationFrame` which
is paused for background tabs. As such the sharks
remain still until the tab is visible again